requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.2
django-allauth==65.1.0
django-axes==7.0.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ipware==7.0.1
django-timezone-field==7.0
django-unfold==0.40.0
django-waffle==4.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
kombu==5.4.2
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-ipware==3.0.0
six==1.16.0
sqlparse==0.5.1
supervisor==4.2.5
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
